Change Interactor to namespaced service object
